requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
Pillow==9.5.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Session==0.5.0
psycopg2-binary==2.9.9
oauthlib==3.2.2
requests-oauthlib==1.3.1
Werkzeug==3.0.1
gunicorn==21.2.0
fal-client==0.1.0 